+++ package.json
@@ -36,7 +36,7 @@
     "start": "vite",
     "build-tests": "esbuild test/index.ts --target=es2020 --bundle --keep-names > test/test-bundle.js",
     "test-ecc": "esbuild ./test/ecc.ts --bundle | npx tapout",
-    "test-rsa": "",
+    "test-rsa": "esbuild ./test/rsa.ts --bundle | npx tapout",
     "test": "npm run build && npm run build-tests && npm run test-tape-run",
     "test-tape-run": "cat test/index.html | tape-run --input=html --static=test | tap-spec",
     "build-cjs": "esbuild src/*.ts --target=es2020 --format=cjs --keep-names --tsconfig=tsconfig.build.json --outdir=./dist --out-extension:.js=.cjs --sourcemap",
